add a task solution
