+++ src/components/TodoFilter/TodoFilter.tsx
@@ -48,12 +48,14 @@ export const TodoFilter: React.FC = () => {
 
         <span className="icon is-right" style={{ pointerEvents: 'all' }}>
           {/* eslint-disable-next-line jsx-a11y/control-has-associated-label */}
-          <button
-            data-cy="clearSearchButton"
-            type="button"
-            className="delete"
-            onClick={() => dispatch(actions.setQuery(''))}
-          />
+          {filter.query.length !== 0 && (
+            <button
+              data-cy="clearSearchButton"
+              type="button"
+              className="delete"
+              onClick={() => dispatch(actions.setQuery(''))}
+            />
+          )}
         </span>
       </p>
     </form>
