+++ src/components/TodoList/TodoList.tsx
@@ -1,127 +1,89 @@
 /* eslint-disable no-console */
 
 //hooks
-import React, { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { useAppSelector } from '../../hooks';
-
-//api
-import { getTodos } from '../../api';
 
 //store
-import { actions as todosActions } from '../../features/todos';
 import { actions as currentTodoActions } from '../../features/currentTodo';
 
 //services
 import classNames from 'classnames';
 
-//components
-import { Loader } from '../Loader';
+//type
+import { Todo } from '../../types/Todo';
 
-export const TodoList: React.FC = () => {
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const todos = useAppSelector(state => state.todos);
-  const filter = useAppSelector(state => state.filter);
+type Props = {
+  todos: Todo[];
+};
+
+export const TodoList: React.FC<Props> = ({ todos }) => {
   const dispatch = useDispatch();
-
-  useEffect(() => {
-    const loadTodos = async () => {
-      try {
-        setIsLoading(true);
-
-        await new Promise(resolve => setTimeout(resolve, 1500));
-
-        const data = await getTodos();
-
-        dispatch(todosActions.set(data));
-      } catch (error) {
-        console.log(error);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    loadTodos();
-  }, [dispatch]);
-
-  const filteredTodos = todos.filter(todo => {
-    const matchesQuery = todo.title
-      .toLowerCase()
-      .includes(filter.query.toLowerCase());
-
-    const matchesStatus =
-      filter.status === 'all' ||
-      (filter.status === 'active' && !todo.completed) ||
-      (filter.status === 'completed' && todo.completed);
-
-    return matchesQuery && matchesStatus;
-  });
 
   return (
     <>
-      {isLoading ? (
-        <Loader />
-      ) : (
-        <>
-          {filteredTodos.length <= 0 && (
-            <p className="notification is-warning">
-              There are no todos matching current filter criteria
-            </p>
-          )}
-          {filteredTodos.length > 0 && (
-            <table className="table is-narrow is-fullwidth">
-              <thead>
-                <tr>
-                  <th>#</th>
+      {todos.length <= 0 && (
+        <p className="notification is-warning">
+          There are no todos matching current filter criteria
+        </p>
+      )}
+      {todos.length > 0 && (
+        <table className="table is-narrow is-fullwidth">
+          <thead>
+            <tr>
+              <th>#</th>
 
-                  <th>
-                    <span className="icon">
-                      <i className="fas fa-check" />
-                    </span>
-                  </th>
+              <th>
+                <span className="icon">
+                  <i className="fas fa-check" />
+                </span>
+              </th>
 
-                  <th>Title</th>
-                  <th> </th>
+              <th>Title</th>
+              <th> </th>
+            </tr>
+          </thead>
+
+          <tbody>
+            {todos.map(todo => {
+              return (
+                <tr data-cy="todo" key={todo.id}>
+                  <td className="is-vcentered">{todo.id}</td>
+                  <td className="is-vcentered">
+                    {todo.completed && (
+                      <span className="icon" data-cy="iconCompleted">
+                        <i className="fas fa-check" />
+                      </span>
+                    )}
+                  </td>
+
+                  <td className="is-vcentered is-expanded">
+                    <p
+                      className={classNames(
+                        { 'has-text-success': todo.completed },
+                        { 'has-text-danger': !todo.completed },
+                      )}
+                    >
+                      {todo.title}
+                    </p>
+                  </td>
+
+                  <td className="has-text-right is-vcentered">
+                    <button
+                      data-cy="selectButton"
+                      className="button"
+                      type="button"
+                      onClick={() => dispatch(currentTodoActions.set(todo))}
+                    >
+                      <span className="icon">
+                        <i className="far fa-eye" />
+                      </span>
+                    </button>
+                  </td>
                 </tr>
-              </thead>
-
-              <tbody>
-                {filteredTodos.map(todo => {
-                  return (
-                    <tr data-cy="todo" key={todo.id}>
-                      <td className="is-vcentered">{todo.id}</td>
-                      <td className="is-vcentered"> </td>
-
-                      <td className="is-vcentered is-expanded">
-                        <p
-                          className={classNames(
-                            { 'has-text-success': todo.completed },
-                            { 'has-text-danger': !todo.completed },
-                          )}
-                        >
-                          {todo.title}
-                        </p>
-                      </td>
-
-                      <td className="has-text-right is-vcentered">
-                        <button
-                          data-cy="selectButton"
-                          className="button"
-                          type="button"
-                          onClick={() => dispatch(currentTodoActions.set(todo))}
-                        >
-                          <span className="icon">
-                            <i className="far fa-eye" />
-                          </span>
-                        </button>
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          )}
-        </>
+              );
+            })}
+          </tbody>
+        </table>
       )}
     </>
   );
